# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.ErrorInfo:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/dto.ValidationDetail'
        type: array
      help:
        type: string
      message:
        type: string
      request_id:
        type: string
      timestamp:
        type: string
    type: object
  dto.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/dto.ErrorInfo'
      meta:
        $ref: '#/definitions/dto.Meta'
      success:
        type: boolean
    type: object
...
  handler.ProductListResponse:
    description: Product list item with basic information
    properties:
      barcode:
        example: "6901234567890"
        type: string
      category_id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      code:
        example: SKU-001
        type: string
      created_at:
        example: "2026-01-24T12:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Sample Product
        type: string
      purchase_price:
        example: 50
        type: number
      selling_price:
        example: 100
        type: number
      sort_order:
        example: 0
        type: integer
      status:
        enum:
        - active
        - inactive
        - discontinued
        example: active
        type: string
      unit:
        example: pcs
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@erp.example.com
    name: API Support
    url: https://github.com/erp/backend
  description: 进销存系统后端 API - 基于 DDD 设计的库存管理系统
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ERP Backend API
  version: "1.0"
paths:
  /catalog/categories/{category_id}/products:
    get:
      consumes:
      - application/json
      description: Retrieve products belonging to a specific category
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Category ID
        format: uuid
        in: path
        name: category_id
        required: true
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      - description: Product status
        enum:
        - active
        - inactive
        - discontinued
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        maximum: 100
        name: page_size
        type: integer
      - default: sort_order
        description: Order by field
        in: query
        name: order_by
        type: string
      - default: asc
        description: Order direction
        enum:
        - asc
        - desc
        in: query
        name: order_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.ProductListResponse'
                  type: array
                meta:
                  $ref: '#/definitions/dto.Meta'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: Get products by category
      tags:
      - products
  /catalog/products:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of products with optional filtering
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Search term (name, code, barcode)
        in: query
        name: search
        type: string
      - description: Product status
        enum:
        - active
        - inactive
        - discontinued
        in: query
        name: status
        type: string
      - description: Category ID
        format: uuid
        in: query
        name: category_id
        type: string
      - description: Unit of measure
        in: query
        name: unit
        type: string
      - description: Minimum selling price
        in: query
        name: min_price
        type: number
      - description: Maximum selling price
        in: query
        name: max_price
        type: number
      - description: Filter by barcode presence
        in: query
        name: has_barcode
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        maximum: 100
        name: page_size
        type: integer
      - default: sort_order
        description: Order by field
        in: query
        name: order_by
        type: string
      - default: asc
        description: Order direction
        enum:
        - asc
        - desc
        in: query
        name: order_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.ProductListResponse'
                  type: array
                meta:
                  $ref: '#/definitions/dto.Meta'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product in the catalog
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Product creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - products
  /catalog/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product by ID
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Retrieve a product by its ID
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: Get product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing product's details
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Product update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - products
  /catalog/products/{id}/activate:
    post:
      consumes:
      - application/json
      description: Activate an inactive product
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: Activate a product
      tags:
      - products
  /catalog/products/{id}/code:
    put:
      consumes:
      - application/json
      description: Update a product's code (SKU)
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: New product code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProductCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: Update product code
      tags:
      - products
  /catalog/products/{id}/deactivate:
    post:
      consumes:
      - application/json
      description: Deactivate an active product
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: Deactivate a product
      tags:
      - products
...
  /system/info:
    get:
      consumes:
      - application/json
      description: Returns basic system information including version and uptime
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.SystemInfoResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      summary: Get system information
      tags:
      - system
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"